pyproject.toml
[tool.poetry]
name = "pokemongame"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
